Implement OSC output. Display log of sent OSC in GUI for testing.
The audio thread now sends an analysis of the audio data to a new OSC
output thread. The OSC output thread sends the last received OSC output
to the designated target installation 60 times per second.

We still need to add a section to the GUI for defining the area of an
installation. This will let us automatically determine which
installation each speaker is associated with based on its location. For
now, Cacophony is just used by default for testing.