Implement "Distance-Based Audio Panning"
This removes the original heuristic-based spatial algorithm in favour of
an implementation of Trond Lossius' "Distance-Based Amplitude Panning"
(2009). The primary advantage of this is much more perceptually accurate
amplitude panning.

Another advantage that DBAP offers over the previous heuristic approach
is that it allows for applying weights to different speakers when
calculating relative speaker amplitudes. This will be useful for
weighting particular speakers for certain sounds. E.g. Some sounds
should only appear in Cosmic Waves, and in turn the speakers within the
cosmic waves installation should have the highest weight.

The implementation is based directly on the algorithm described in the
paper itself and not ported from any prior implementation. The paper can
be found at #40, along with a link to an evaluation by Microsoft where
they list the advantages of DBAP over VBAP and ambisonics.

Closes #40.