Add Installation Editor menu for dynamically updating OSC outputs
A new installation editor sidebar menu has been added which allows for
specifying the target IP, port and OSC address for sending musical data
to each of the installations in the exhibition.

When editing the IP:PORT string the text box will glow red if the string
is not a valid IPV4 address and port. It will disallow updating the
target until the string represents a valid IPV4 address and port. The
target can be updated by pressing `Enter` within either of the text
boxes.

Next on the list is adding the option for assigning one or more
installations to each speaker channel, each with a unique weight.