Add `Master` panel to sidebar GUI. Includes master volume, latency ms.
The "Master Volume" controls the overall volume for the entire exhibition.
It is applied at the end of the audio output process.

The "Realtime Source Latency" is the latency used to pad the audio input
stream for syncing with the audio output stream. This can be very low in
the case that the audio clock is synchronised between input and output,
but will have to be quite high on devices with unsyncronised clocks that
drift significantly.

Closes #61.
Closes #33.