Create a more flexible `source::Signal` type to allow for file seeking
This is necessary in order to allow for a `Continuous` playback mode for
WAV sources. A `Continuous` playback mode plays back WAVs as though they
are constantly playing on an invisible timeline.

In order to ensure sample-accurate continuous playback for WAVs, we will
use this new `seek` method to seek to the exact sample from which the
WAV should resume playing every time a buffer is requested.