Implement Soundscape generation (movement not yet added)
This applies all previously added constraints to the generation of
sound playback on the soundscape thread. The constraints considered are:

- Installation and group assignments for sources.
- Minimum and maximum number of simultaneous sounds per installation,
group and source.
- The rate at which sounds may occur per group and per source.
- The attack, release and playback duration ranges of the source used to
spawn a new sound.

Sounds are currently spawned at a random location within the
installation, chosen via the centroid of all speakers assigned to the
installation and the bounding rectangle around all speaker locations.