Add `Ngon` movement. Add GUI control for `Agent` movement. FFT speedup.
This implements the `Ngon` movement pattern which traces the vertices of
a virtual n-sided polygon that is laid out over an installation. This
movement kind has a number of configurable parameters:

- The speed (in metres per second) at which the sound travels through
the space.
- The number of vertices (and in turn, the number of sides).
- The number of vertices to skip at a time when travelling around the
Ngon. That is, if the ngon has `6` vertices and a skip value of `2`,
sounds will travel in a triangular pattern around the installation.
- The normalised dimensions of the Ngon. These represent the fraction of
the total bounds of the installation over which a sound is able to
travel. E.g. if a normalised width of `0.5` is provided, the ngon
pattern will only reach across half of the width of the installation and
in turn the sounds' movement will be constrained to that area.

**Note** that GUI parameters are not yet provided for these, but will be
provided in a following PR.

GUI parameters have however been added for the `Agent` movement type's
"max speed" and "max force".

This PR also implements a significant speedup for the FFT calculations
by re-using the same FFT planner between all FFT instances.

This PR is related to the continued development on #28.