Implement `Fixed` position and `Ngon` movement GUI.
It is now possible to specify a fixed position within the location via
an XY pad when the "FIXED" movement option is enabled. This position is
a normalised representation of the assigned installation's bounding
rectangle.

It is also now possible to control the the `Ngon` tracing movement
constraints. Please see my comment in the previous commit for a brief
explanation of what these constraints are and how they behave.

Closes #28.