Fix bug where "Simultaneous Sounds" slider would disappear
Currently the "Simultaneous Sounds" slider used within the "Installation
Editor" is accidentally shared with the same slider in the "Soundscape
Editor" panel. This would cause the Installation Editor's slider to
disappear whenever the Soundscape Editor panel was open.

This PR fixes this issue so that both panels use their own unique
sliders.

Closes #92.