Amplify noise walk to reach min and max simultaneous sounds
The soundscape currently uses a "noise walk" to create a "target" number
of sounds for each installation by oscillating between their specified
minimum and maximum simultaneous sounds over time at a very low
frequency. One issue with the noise walk is that it rarely reaches the
-1.0 or 1.0 exactly. This commit amplifies the noisewalk slightly to
reach closer to these bounds, but clamps the values to ensure it does
not exceed them.