Loop at 60FPS rather than waiting for events. Improves macos performance
Currently rather than looping constantly, the audio server stops looping
if there are no active animations until the next user input is received.
This is often a great energy-saving exercise for gui apps, however the
audio server will almost always be animating due to the sounds that are
playing back and in turn there is little benefit of having this mode
enabled. This should speed up the macos build significantly as the
method used to "wake up" the app event loop is surprisingly expensive
and shows up as one of the primary bottlenecks.