Fix bug where some sidebar panels would not allow swipe/mouse scrolling.
This bug particularly causes a lot of frustration when editing many
sources as the user has to move the mouse to the scrollbar and manually
click and drag it in order to navigate up and down the area. This commit
fixes swipe/mouse scrolling behaviour so that it should work in all
side-bar GUI areas.