Add support for radial playback progress bars to custom sound widget.
This is an optional parameter that can display the playback progress of
a sound as a radial progress bar starting at the top and rotating
clockwise as the sound completes.

This is not yet plugged into the sound's actual progress, but is now
ready for when this information is available.