Implement Custom Installations (Add, Remove, Rename)
This refactors the codebase to support custom installations.

Installations now support:

- Adding new isntallations via the "+" button
- Removing installations via the "x" button
- Editing installation names by editing in the textbox and pressing
`Enter`.

Note that when editing and entering a new installation name, the OSC
address will be updated to be the "slugified" (computer-friendly)
version of that name.

By default (if there is no existing state.json file) the audio server
will generate the known beyond perception installations.

The server still supports loading the original fixed installation
states, so it should be fine to use old projects with this version.