Fixes a bug where simultaneous sound soundscape constraints are violated
Previously the soundscape was incorrectly constraining group and source
numbers based only on the number of sounds for each installation as it
added new sounds for those installations. However, the purpose of the
constraints are to constrain groups and sources based on the *global*
number of occurrences of those sounds. This commit addresses this issue
by ensuring the global number of sounds are considered instead.

Closes #120.