Fixes bug where adding an installation could cause audio thread to panic
This resulted in not being able to edit any of the GUI parameters for
thew newly created installation. The root of the cause was that, when
a new installation was created, the speaker count was reduced to 0. This
would cause the DBAP calculation to `panic!` as there must be at least
one speaker for the calculations to make any sense.

Closes #145.