Avoid calculating FFT if installation has no OSC receiving computers
FFT has been showing up as the primary bottleneck on the audio thread.
This commit ensures that, rather than calculating the FFT for each
installation, it is only calculated for those installations that have
one or more computers prepared to receive the audio data via OSC.

This should be particularly useful for the Beyond Perception exhibition
as it seems only the Projection Mapping installation will receive and
visualise the audio data.

This also updates the default number of computers for each installation
to 0 (other than Energetic Vibrations - Projection Mapping). The
`state.json` for the main test project has been updated accordingly.