Fix bug where sounds would glitch if they panned too quickly
Previously the gain for each speaker for each channel in each sound was
only re-calculated once per buffer. This meant that if a sound was
moving fast enough over the duration of an audio buffer  to cause a
significant change in DBAP gain, a clipping effect could occur.

This commit fixes this by ensuring that the buffer linearly interpolates
the gain from the gain at the end of the previous buffer to the newly
calculated gain.

Closes #155.