Ensure that program closes properly if thread communication fails
Currently it is possible for a thread to die and for the program to
continue running, despite not being able to successfully communicate
between threads. This can cause seemingly unrelated strange behaviour in
the GUI.

This commit makes sure that the program closes immediately with an error
message if a message cannot be sent across threads due to one of the
threads dying.

cc @freesig this should also make sure the program doesn't continue
running in a broken state if one of the threads panic and should allow
a keepalive script to immediately restart the server.