Only calculate DBAP for speakers within `audio::PROXIMITY_LIMIT`
This changes the DBAP loop to skip speakers that are not within
proximity of the sound channel for which the calculation is occurring.
This should reduce the amount of time spent performing DBAP to a
fraction of the original cost.

Also adds `rayon` and `crossbeam` to the crate just in case we need
them.