Fix bug where GUI would allow for specifying NGON with `0` sides
This fixes an issue where the source editor GUI would allow the user to
select `0` as the number of sides of the equilateral NGON. This would
cause the program to `panic!` when the value was used as the divisor in
a remainder calculation a little later on. This has been addressed by
bounding the slider to a minimum number of sides of "1" (in which case
the sound will stay centred at the NGON's position and not move - this
behaviour is unchanged).

Closes #188.